requirements.txt
Flask
SpeechRecognition
gtts
pydub
google-generativeai
python-dotenv